Use request ID as task description (#1451)
Currently task ID is used as a key in the map. This is OK while there
is only one session because task ID is unique inside the session.
We are creating separate session for requests with proxy so task ID
is not unique anymore. Request ID is unique across the Network instance
and task description is available same a the task ID.

Relates-To: OLPEDGE-2846

Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>