Fix parsing of server DateTime (#1499)
std::get_time does not claim support of time zones in the string
parameter, and seems to be subject of locale set on the system. An
example is a following result on linux:

ParseTime(Thu, 1 Jan 1970 0:00:00 GMT)=-2209075200

::strptime is also a subject to the time zones of locale, but it also
claim support for parsing a time zone from the passed string. Thus,
additional logic incorporated to offset parsed DataTime to the local
timezone.

Relates-To: IOTSDK-24021

Signed-off-by: Mykhailo Diachenko <ext-mykhailo.z.diachenko@here.com>