Add per-thread context values to logging system
This change adds a per-thread logging context to the logging
system, which can be set by the user and will then be logged
from all threads and tasks created with that context.

This is useful to correlate requests from the client to any
network requests made based on that request.

The logging formatter must be set up by including a
`MessageFormatter::ElementType::ContextValue` with the 'format'
set to the context key to use.

Resolves: OLPEDGE-2893
Signed-off-by: Hertlein, Ulrich <ulrich.hertlein@here.com>