Background subscriber to check current value (#1543)
To prevent triggering handlers several times in a row. Currently
this should not harm the app because NSURLSessionDownloadTask tasks
are skipped in restartCurrentTasks.

Testing on iOS-13 simulator it was possible to get two same states
in a row using notifications. Maybe a simulator issue or iOS-13 case
which is the first where background mode delegates are in the Scene
delegate instead of Application delegate.

Relates-To: DATASDK-39

Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>