Custom body to be sent to URL from Credentials
Here credentials file specifies the endpoint that
should be used to retrieve authentication token.
HNAV backend, which implements mTLS authentication
also conforms to a bit different body and response
schema.

Add a possibility to add a custom body for the authentication request
and parsing of token from HNAV mTLS server.

Relates-To: HERESDK-7942
Signed-off-by: Mykhailo Diachenko <ext-mykhailo.z.diachenko@here.com>