Use porting::optional in olp-cpp-sdk-write instead of boost::optional. (#1626)
* Use std::optional if available.

Currently, SDK requires C++11 minimum.
So, boost::optional type is used for optional values.
For C++17 and above more convenient is to use std::optional instead.
The task NLAM-23 is about making this type configurable.
This commit is a second part of the task: olp-cpp-sdk-write.

Relates-To: NLAM-23
Signed-off-by: sopov <ext-alexander.sopov@here.com>

* Use std::optional if available.

Currently, SDK requires C++11 minimum.
So, boost::optional type is used for optional values.
For C++17 and above more convenient is to use std::optional instead.
The task NLAM-23 is about making this type configurable.
This commit is a second part of the task: olp-cpp-sdk-write.

Relates-To: NLAM-23
Signed-off-by: sopov <ext-alexander.sopov@here.com>

---------

Signed-off-by: sopov <ext-alexander.sopov@here.com>