Remove custom certificate handling (#1635)
Certificate validation logic implemented in the
Data SDK is a subject of transient errors, as
it is implemented on top of deprecated APIs.

Given that there's no handling of the custom
certificates and logic corresponds to the
default certificate handling in the iOS, custom
implementation can be removed in the faviour
of default handling.

Relates-To: HERESDK-5806

Signed-off-by: Mykhailo Diachenko <ext-mykhailo.z.diachenko@here.com>