Add censor API to Log (#1639)
* Add censor API to Log

It allows to set values to be masked from the log messages at runtime

Relates-To: NLAM-140
Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>

* Extend Logs tests caused by formatting changes

Codecov prefers to have patch config to be not lower than the rest
of repository. The file has not been formatted properly before
so there are changes not related to the censoring

Relates-To: NLAM-140
Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>

* Add removing censor API to Log

There are secrets that can be updated with time and it make sense
not to censor them anymore not to let performance suffer.

Relates-To: NLAM-140
Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>

---------

Signed-off-by: Rustam Gamidov <ext-rustam.gamidov@here.com>